+++ moddb/__init__.py
@@ -2,7 +2,7 @@ from .base import search, parse_page, parse_results, login, logout, front_page, 
 from .utils import LOGGER, BASE_URL, soup, request, get_page, Object
 from .enums import *
 from .pages import *
-from .client import Client
+from .client import Client, Thread
 
 import requests